# sprint14_deliverables/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Benchmark API",
    "description": "Programmatic API for LLM benchmarking, routing, and analytics",
    "contact": {
      "name": "AI-Benchmark Team"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.1.0"
  },
  "paths": {
...
    "/api/v1/benchmarks/run": {
      "post": {
        "tags": [
          "benchmarks"
        ],
        "summary": "Run Benchmark",
        "operationId": "run_benchmark_api_v1_benchmarks_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/benchmarks/results/{result_id}": {
      "get": {
        "tags": [
          "benchmarks"
        ],
        "summary": "Get Result",
        "operationId": "get_result_api_v1_benchmarks_results__result_id__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Result Api V1 Benchmarks Results  Result Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/generate": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Reports",
        "operationId": "generate_reports_api_v1_reports_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Report Api V1 Reports  Report Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchmarkResultResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "overall": {
            "type": "number",
            "title": "Overall"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkScoreSchema"
            },
            "type": "array",
            "title": "Scores"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider",
          "model",
          "overall"
        ],
        "title": "BenchmarkResultResponse"
      },
      "BenchmarkRunRequest": {
        "properties": {
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "benchmarks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmarks"
          },
          "messages": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messages"
          },
          "fallback_enabled": {
            "type": "boolean",
            "title": "Fallback Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "provider_name",
          "model"
        ],
        "title": "BenchmarkRunRequest"
      },
      "BenchmarkScoreSchema": {
        "properties": {
          "benchmark": {
            "type": "string",
            "title": "Benchmark"
          },
          "raw": {
            "type": "number",
            "title": "Raw"
          },
          "normalized": {
            "type": "number",
            "title": "Normalized"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "weighted": {
            "type": "number",
            "title": "Weighted"
          }
        },
        "type": "object",
        "required": [
          "benchmark",
          "raw",
          "normalized",
          "weight",
          "weighted"
        ],
        "title": "BenchmarkScoreSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportGenerateRequest": {
        "properties": {
          "result_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Ids"
          },
          "formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Formats"
          },
          "out_dir": {
            "type": "string",
            "title": "Out Dir",
            "default": "reports"
          }
        },
        "type": "object",
        "title": "ReportGenerateRequest"
      },
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Formats"
          },
          "files": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Files"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "ReportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}